pyproject.toml
[project]
name = "ml-template"
version = "0.1.0"
description = "Vishvam's core ML starter template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "catppuccin>=2.5.0",
  "gradio>=6.14.0",
  "matplotlib>=3.10.9",
  "numpy>=2.4.4",
  "pandas>=3.0.2",
  "python-dotenv>=1.2.2",
  "pyyaml>=6.0.3",
  "rich>=15.0.0",
  "safetensors>=0.7.0",
  "scikit-learn>=1.8.0",
  "scipy>=1.17.1",
  "tensorboard>=2.20.0",
  "torch>=2.11.0",
  "torchvision>=0.26.0",
  "tqdm>=4.67.3",
  "typer>=0.25.1",
]
[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
  "*_pb2*.py*",
]

line-length = 80
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
exclude = ["*_pb2*.py"]
select = ["E4", "E7", "E9", "F"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = false
max-doc-length = 80

[tool.ruff.format]
exclude = ["*_pb2*.py"]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
branch = true
omit = ["*_pb2*.py", "tests/*", "*/tests/*", "*_test.py", "test_*.py"]

[tool.coverage.report]
skip_covered = true
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[dependency-groups]
dev = ["pytest>=9.0.3", "ruff>=0.15.12"]
